# specs/001-haruhanjang-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 하루한장 API
  version: 0.1.0
  description: |
    사장님 PWA(S1~S8)와 기관 대시보드(S7)가 쓰는 REST API.
    근거: spec.md FR-xxx, design/SD_01 IPO, SD_02 화면, SD_03 테이블·뷰.

    공통 규칙
    - 모든 경로는 `/api` 아래. 인증은 HttpOnly 쿠키(`hrh_session`, 서명 JWT).
    - 날짜는 KST 기준 `YYYY-MM-DD`.
    - 게이트 차단은 HTTP 409 + `GateBlocked` 본문. 화면은 `gate` 코드로 SD_02 §11-2 표준 문구를 고른다.
    - 사장님 API는 로그인한 계정의 가게 데이터만 다룬다(FR-091). 경로에 store_id를 받지 않는다.
    - 기관 API(`/api/org/*`)는 익명 집계 뷰만 조회한다(FR-076, research R7).
    - 인가(RBAC): 각 경로의 권한은 data-model §7-2. 권한이 없으면 403 Forbidden, 게이트 차단은 409.
servers:
  - url: /api
...
paths:
...
  /store/closed-days:
    put:
      tags: [store]
      summary: 쉬는 요일 입력 (UC1 A2, UC3 A1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [daysOfWeek]
              properties:
                daysOfWeek: { type: array, items: { type: integer, minimum: 1, maximum: 7 }, description: '1=월 … 7=일. 빈 배열 = 쉬는 날 없음' }
      responses:
        '204': { description: '저장' }
  /store/alert-settings:
    patch:
      tags: [store]
      summary: 경보 수신 켜기·끄기 (UC6 A1, FR-053)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [pushEnabled], properties: { pushEnabled: { type: boolean } } }
      responses:
        '204': { description: '저장' }
  /consents:
    post:
      tags: [store]
      summary: 동의 사건 추가 (익명 참여 동의·철회, FR-017, FR-064)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consentItemCode, agreed]
              properties:
                consentItemCode: { type: string, enum: [anon_stats] }
                agreed: { type: boolean }
      responses:
        '201': { description: '기록' }
...
  /records:
    get:
      tags: [records]
      summary: 기간 기록 + 외부 환경 결합 (달력·기기 내 분석 입력, P3·P4)
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date } }
        - { in: query, name: to, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/DailyRecord' } }
  /records/{date}:
    parameters:
      - { in: path, name: date, required: true, schema: { type: string, format: date } }
    get:
      tags: [records]
      summary: 그날의 한 장 (S4 4.3)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyRecord' }
        '404': { description: '기록 안 한 날' }
    put:
      tags: [records]
      summary: 기록 저장·갱신 (P2 2.5~2.6, UC2 A2·A3)
      description: |
        FR-020~029. 같은 날짜는 갱신(revision+1). 미래 날짜는 422.
        외부 데이터 결합은 응답 후 비동기로 수행하며, 응답의 env 는 결합 전이면 needs_check.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordInput' }
      responses:
        '200':
          description: 저장됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyRecord' }
        '409': { $ref: '#/components/responses/GateBlocked' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /records/months/{month}/summary:
    get:
      tags: [records]
      summary: 월간 요약 (S4 4.4, v_store_month_summary)
      parameters:
        - { in: path, name: month, required: true, schema: { type: string, pattern: '^\d{4}-\d{2}$' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  recordedDays: { type: integer }
                  goodDays: { type: integer }
                  normalDays: { type: integer }
                  badDays: { type: integer }
                  topEvents: { type: array, items: { $ref: '#/components/schemas/CodeCount' } }
...
components:
  responses:
...
    ValidationError:
      description: 입력 형식 오류 (화면에는 기술 용어 없이 표시, FR-083)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: validation }
              fields: { type: object, additionalProperties: { type: string } }
    GateBlocked:
      description: 게이트 차단 (SD_02 §11, SD_03 §13)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GateBlocked' }
  schemas:
...
    GateBlocked:
      type: object
      required: [gate, reasonKey, releasedBy]
      properties:
        gate: { type: string, enum: [G0, G1, G2, G3, G4, G5, G6, G7, G8] }
        reasonKey: { type: string, description: 'SD_02 §11-2 표준 문구 키 (예: g3.reason)' }
        releasedBy: { type: string, enum: [self, others], description: 'others 면 화면에 해제 버튼을 두지 않는다 (FR-002)' }
        detail:
          type: object
          description: '문구 치환 값 (예: G3 recordCount, minRecords). 가게 수 등 익명성 수치는 넣지 않는다'
          additionalProperties: true
...
    CodeCount:
      type: object
      properties:
        code: { type: string }
        count: { type: integer }
...
    RecordInput:
      type: object
      required: [dayMood, customerLevel, eventTypeCodes]
      properties:
        dayMood: { type: string, enum: [good, normal, bad] }
        customerLevel: { type: string, enum: [many, usual, few] }
        eventTypeCodes:
          type: array
          uniqueItems: true
          items: { type: string, enum: [rain, discount, new_menu, sns_post, group_guest, stock_out, staff_absent] }
        salesBandCode: { type: string, nullable: true }
        clientQueuedAt: { type: string, format: date-time, nullable: true, description: 오프라인 대기열에서 보낸 경우 }
    EnvStatus:
      type: object
      properties:
        weatherCode: { type: string, nullable: true }
        weatherStatus: { type: string, enum: [ok, needs_check] }
        isHoliday: { type: boolean, nullable: true }
        holidayStatus: { type: string, enum: [ok, needs_check] }
        localEventName: { type: string, nullable: true }
        localEventStatus: { type: string, enum: [ok, needs_check] }
    DailyRecord:
      type: object
      properties:
        recordDate: { type: string, format: date }
        dayMood: { type: string, enum: [good, normal, bad] }
        customerLevel: { type: string, enum: [many, usual, few] }
        eventTypeCodes: { type: array, items: { type: string } }
        salesBandCode: { type: string, nullable: true }
        dayOfWeek: { type: integer, minimum: 1, maximum: 7 }
        revision: { type: integer }
        env: { $ref: '#/components/schemas/EnvStatus' }
...